+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Features
+
+- Repeat image layer 'Sprite' based upon Tiled repeat x / repeat y settings (#123)
+
 ## v0.9.1
 
 ### Features